Cria exemplo de navegação no app expo-example
