+++ expo-example/package.json
@@ -9,7 +9,8 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@shapeddev/shaped-expo-plugin": "0.0.3",
+    "@react-navigation/native": "^7.0.18",
+    "@react-navigation/native-stack": "^7.3.2",
     "@shopify/react-native-skia": "1.11.7",
     "expo": "~52.0.37",
     "expo-dev-client": "~5.0.12",
@@ -17,7 +18,10 @@
     "react": "18.3.1",
     "react-native": "0.76.7",
     "react-native-reanimated": "~3.16.1",
-    "react-native-svg": "15.8.0"
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.4.0",
+    "react-native-svg": "15.8.0",
+    "expo-build-properties": "~0.13.2"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
